fix(application): resolve bugs in schedule and repository logic
- Fixed null reference issue in Appointment handler
- Corrected repository query to return active subscriptions only
- Resolved DI registration error causing missing service injection
- Fixed mapper configuration in controllers to properly map DTOs
- Removed obsolete commented code for cleaner implementation
- Verified integration with Branch and Staff modules after fixes

+++ Viora/Viora.Domain/Plans/PlanLimitedFeature.cs
@@ -7,10 +7,10 @@ public class PlanLimitedFeature : Entity
 {
     public Guid PlanId { get; private set; }
     public Guid LimitedFeatureId { get; private set; }
-    public int LimitValue { get; private set; }
+    public long LimitValue { get; private set; }
 
     public LimitedFeature LimitedFeature { get; private set; }
-    private PlanLimitedFeature(Guid id, Guid planId, Guid limitedFeatureId, int limitValue) : base(id)
+    private PlanLimitedFeature(Guid id, Guid planId, Guid limitedFeatureId, long limitValue) : base(id)
     {
         PlanId = planId;
         LimitedFeatureId = limitedFeatureId;
